% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ro_pkgs}
\alias{ro_pkgs}
\title{List ropensci packages}
\usage{
ro_pkgs(query = NULL, category = NULL, on_cran = NULL, simplify = FALSE,
  ...)
}
\arguments{
\item{...}{Named options passed on to \code{\link[devtools]{install}}}
}
\description{
List ropensci packages
}
\examples{
\donttest{
# list all packages
ro_pkgs()

# query for a pacakage
ro_pkgs(category = "history")
ro_pkgs(category = "archeology")
ro_pkgs(category = "economics")
ro_pkgs(on_cran = TRUE)
ro_pkgs(query = "species")
ro_pkgs(query = "alfjafaksd")
}
}

