% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_ropensci.R
\name{install_ropensci}
\alias{install_ropensci}
\title{Install ropensci packages}
\usage{
install_ropensci(pkgs, type = "source", ...)
}
\arguments{
\item{pkgs}{One or more packages to install}

\item{type}{character; Only source for now, but hopefully soon support binaries.}

\item{...}{Named options passed on to \code{\link[devtools]{install}}}
}
\description{
Install ropensci packages
}
\details{
Only installs source (non-binary) packages from GitHub for now.
}
\examples{
\donttest{
install_ropensci(pkgs = 'musemeta')
install_ropensci(pkgs = 'musemeta', quiet = TRUE)
install_ropensci(pkgs = c('musemeta','rerddap'))
install_ropensci(pkgs = 'alm', ref = "v0.3.0")

# not found
install_ropensci(pkgs = 'stuffthings')
}
}

