% Generated by roxygen2 (4.0.2): do not edit by hand
\name{on_cran}
\alias{on_cran}
\title{Check if an ropensci package is on CRAN.}
\usage{
on_cran(pkgs)
}
\arguments{
\item{pkgs}{One or more package names}
}
\description{
Check if an ropensci package is on CRAN.
}
\examples{
\donttest{
on_cran('taxize')
on_cran('musemeta')
on_cran(pkgs=c('taxize','musemeta'))
}
}

