% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_cran.R
\name{on_cran}
\alias{on_cran}
\title{Check if an ropensci package is on CRAN.}
\usage{
on_cran(pkgs, force_update = TRUE)
}
\arguments{
\item{pkgs}{One or more package names}

\item{force_update}{(logical) force update? Default: `TRUE`}
}
\description{
Check if an ropensci package is on CRAN.
}
\examples{
\donttest{
on_cran(pkgs='taxize')
on_cran('musemeta')
on_cran(pkgs=c('taxize','musemeta'))
}
}
