% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_pkgs.R
\name{ro_pkgs}
\alias{ro_pkgs}
\title{List ropensci packages}
\usage{
ro_pkgs(query = NULL, category = NULL, on_cran = NULL,
  maintainer = NULL, simplify = FALSE, ...)
}
\arguments{
\item{query}{Query terms, uses \code{grep}}

\item{category}{Category to get, matches exactly}

\item{on_cran}{(logical) Whether to select only those on CRAN, default: all}

\item{maintainer}{(character) A maintainer name}

\item{simplify}{(logical) Simplify output to more important fields}

\item{...}{Named options passed on to \code{\link[devtools]{install}}}
}
\description{
List ropensci packages
}
\examples{
\donttest{
# list all packages
ro_pkgs()

# query for a pacakage
ro_pkgs(category = "history")
ro_pkgs(category = "archeology")
ro_pkgs(category = "economics")
ro_pkgs(maintainer = "Scott Chamberlain")
ro_pkgs(on_cran = TRUE)
ro_pkgs(query = "species")
ro_pkgs(query = "alfjafaksd")
}
}
