% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropkgs-package.R
\docType{package}
\name{ropkgs-package}
\alias{ropkgs-package}
\alias{ropkgs}
\title{Install rOpenSci packages}
\description{
Install rOpenSci packages
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
